moved unit test under their own subdirectory
